Break App component out from router logic/config
